Player receives damage
Copy the child, "Enemy Canvas," of an enemy game object and paste it into the Player game object. Be sure the child is being copied as well. Rename it to "Player Canvas." Drag the prefab, CBT, into the PlayerReceivesDamageScript of the "CBT Prefab" slot. Be sure you're dragging the prefab CBT into the slot and not the game object CBT (big difference!) Re-position the transform of the player child CBT game object to (0,0,0) so that it is now aligned properly with the player game object. Drag the "player" gameobject into the PlayerReceivesDamageScript as well.